Creating components for page.tsx
